+++ src/components/index.ts
@@ -4,10 +4,11 @@ import { PersonCard } from "./PersonCard";
 import { ListCard } from "./ListCard";
 import { SeriesCard } from "./SeriesCard";
 import { SearchBar } from "./SearchBar";
+import { SortSelect } from "./SortSelect";
 import { GlobalSearch } from "./GlobalSearch";
 import { HomepageVisual } from "./HomepageVisual";
 import { SafeImage } from "./SafeImage";
 import { Breadcrumbs } from "./Breadcrumbs";
 import { BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState } from "./Skeletons";
 
-export { SectionHeading, BookCard, PersonCard, ListCard, SeriesCard, SearchBar, GlobalSearch, HomepageVisual, SafeImage, Breadcrumbs, BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState };
+export { SectionHeading, BookCard, PersonCard, ListCard, SeriesCard, SearchBar, SortSelect, GlobalSearch, HomepageVisual, SafeImage, Breadcrumbs, BookCardSkeleton, PersonCardSkeleton, ListCardSkeleton, EmptyState };